+++ src/app/core/components/redeem-gift/redeem-gift.component.html
@@ -1,0 +1,33 @@
+<!-- @format -->
+<div class="panel-title">Redeem Gift</div>
+<p>
+  If you've been given a gift code, you can redeem it for free storage below.
+</p>
+<form
+  [formGroup]="promoForm"
+  class="dialog-form"
+  (submit)="onPromoFormSubmit(promoForm.value)"
+>
+  <div class="dialog-form-field">
+    <label for="promoCode">Gift code</label>
+    <input
+      type="text"
+      class="form-control"
+      id="promoCode"
+      name="name"
+      formControlName="code"
+    />
+  </div>
+  <div class="dialog-form-field">
+    <button class="btn btn-primary" [disabled]="promoForm.invalid || waiting">
+      Redeem
+    </button>
+  </div>
+  <div
+    class="results-message"
+    *ngIf="resultMessage"
+    [class.successful]="resultMessage?.successful"
+  >
+    {{ resultMessage.message | prConstants }}
+  </div>
+</form>
